Not finished by any means...
Implemented a progress bar and it now does as much as the sql query that most people are happy with.

Will add computer groups and users soon. Only mobile devices work for now.

Debating on if this should be integrated into the original scope report or if it will end up being a stand alone.